#1992 - Separated out components inside HeaderCell.
