+++ src/components/Table/components/HeaderCell/ResizableHeaderCell.jsx
@@ -1,0 +1,31 @@
+import React from "react";
+
+import { Resizable } from "react-resizable";
+
+import { CellContent } from ".";
+
+const ResizableHeaderCell = props => {
+  const { onResize, width, onResizeStop, ...restProps } = props;
+
+  if (!width) {
+    return <CellContent {...restProps} />;
+  }
+
+  return (
+    <Resizable
+      {...{ onResize, onResizeStop, width }}
+      draggableOpts={{ enableUserSelectHack: false }}
+      height={0}
+      handle={
+        <span
+          className="react-resizable-handle"
+          onClick={e => e.stopPropagation()}
+        />
+      }
+    >
+      <CellContent {...restProps} />
+    </Resizable>
+  );
+};
+
+export default ResizableHeaderCell;
